+++ src/drm/drm_type.cpp
@@ -1,0 +1,21 @@
+#include "drm_type.h"
+
+bool DRMType::getIsInstalled()
+{
+    return this->isInstalled;
+}
+
+void DRMType::setIsInstalled()
+{
+    this->isInstalled = true;
+}
+
+QDir DRMType::getRootDir()
+{
+    return this->rootDir;
+}
+
+void DRMType::setRootDir(const QDir& rootDir)
+{
+    this->rootDir = rootDir;
+}
