+++ src/drm/origin_drm.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "drm_type.h"
+
+class OriginDRM : public DRMType
+{
+    void checkOriginExists();
+};
