+++ src/drm/steam_drm.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "drm_type.h"
+
+class SteamDRM : public DRMType
+{
+    void checkSteamExists();
+};
+
