+++ src/drm/uplay_drm.cpp
@@ -1,0 +1,1 @@
+#include "uplay_drm.h"
