+++ src/drm/uplay_drm.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "drm_type.h"
+
+class UplayDRM : public DRMType
+{
+    void checkUplayExists();
+};
+
+
